refactor(request): extract branch update cooldown tracking
